requirements.txt
streamlit
sentence-transformers
datasets
faiss-cpu
numpy
